ebos: add a simple demonstration of the solvent module
it just injects one mole per second per cubic meter into the first
hundred cells.

To be done: output and the disabled assertation in the well model
needs to be fixed properly. (For both I'm a bit unsure how to do it.)

note that flow_ebos does not yet compile with this because its well
model currently hard-codes the number of PDEs.