blackoil solvent: allow the upwind cell not to be the central one of the stencil
i.e., its values should be seen as constants. also, the loop over the
phases for the storage term is now done by the solvent module instead
of the calling code. (this allows for better flexibility and in the
case of the storage term it does not negatively affect performance.)